+++ harjoitustyo/src/database_connection.py
@@ -3,16 +3,16 @@ import sqlite3
 
 dirname = os.path.dirname(__file__)
 
-connection = sqlite3.connect("database.db")
-connection.row_factory = sqlite3.Row
+db_connection = sqlite3.connect("database.db")
+db_connection.row_factory = sqlite3.Row
 
 
 def get_database_connection():
-    return connection
+    return db_connection
 
 
 def test_connection():
-    connection = sqlite3.connect("tests.db")
-    connection.row_factory = sqlite3.Row
+    db_connection = sqlite3.connect("tests.db")
+    db_connection.row_factory = sqlite3.Row
 
-    return connection
+    return db_connection