+++ harjoitustyo/src/services/services.py
@@ -10,11 +10,14 @@ class UsernameInUseError(Exception):
 class InvalidGradeError(Exception):
     pass
 
+class NegativeEctsError(Exception):
+    pass
+
 
 class SisuService:
 
     def __init__(self):
-        
+
         self.user_repository = user_repository
         self.course_repository = course_repository
 
@@ -49,7 +52,7 @@ class SisuService:
         course_repository.create_mandatory_enrollments_id(user_id[0], degree_id)
 
 
-    def calculate_GPA(self, user_id):
+    def calculate_gpa(self, user_id):
 
         user_gpa = course_repository.calculate_GPA(user_id)
 
@@ -83,14 +86,22 @@ class SisuService:
         course_repository.update_grade(course_id, user_id, grade)
 
 
-    def add_course_to_curriculum(course_name, ects, degree_id, user_id):
+    def add_course_to_curriculum(self, course_name, ects, degree_id, user_id):
 
-        course_repository.add_course_to_curriculum(ects, degree_id, user_id)
+        ret_value = course_repository.add_course_to_curriculum(course_name, ects, degree_id, user_id)
+
+        if ret_value == -1:
+            raise NegativeEctsError
 
 
     def delete_enrollment(self, user_id, del_id):
 
         course_repository.delete_enrollment(user_id, del_id)
+
+
+    def users_size(self):
+
+        return user_repository.users_size()
 
 
 sisu_service = SisuService()